agent-worker/requirements.txt
anthropic==0.40.0
PyGithub==2.4.0
GitPython==3.1.43
redis==5.0.8
httpx==0.27.2
PyYAML==6.0.2
python-dotenv==1.0.1
PyJWT==2.9.0
cryptography==43.0.3
